rewrote controllers as classes (static for now)
