+++ src/controllers/homepageController.js
@@ -1,7 +1,9 @@
 const { join } = require('path');
 
-const homepageController = (req, res) => {
-    res.sendFile(join(__dirname, '../public/subscribe.html'));
+class HomepageController {
+    static getStaticHomepage(req, res) {
+        res.sendFile(join(__dirname, '../public/subscribe.html'));
+    }
 }
 
-module.exports = { homepageController };
+module.exports = HomepageController;
